Remove duplicate and unnecessary otlp exporter wrapper
