+++ exporter/otlpexporter/otlp.go
@@ -1,10 +1,10 @@
-// Copyright The OpenTelemetry Authors
+// Copyright  OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,146 +16,81 @@ package otlpexporter
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"sync"
+	"time"
 
-	"go.opentelemetry.io/collector/component"
+	"github.com/cenkalti/backoff"
+	"google.golang.org/genproto/googleapis/rpc/errdetails"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/metadata"
+	"google.golang.org/grpc/status"
+
+	"go.opentelemetry.io/collector/component/componenterror"
 	"go.opentelemetry.io/collector/config/configmodels"
 	"go.opentelemetry.io/collector/consumer/pdata"
 	"go.opentelemetry.io/collector/consumer/pdatautil"
-	"go.opentelemetry.io/collector/exporter/exporterhelper"
 	"go.opentelemetry.io/collector/internal/data"
 	otlpmetrics "go.opentelemetry.io/collector/internal/data/opentelemetry-proto-gen/collector/metrics/v1"
 	otlptrace "go.opentelemetry.io/collector/internal/data/opentelemetry-proto-gen/collector/trace/v1"
 	otlplogs "go.opentelemetry.io/collector/internal/data/opentelemetry-proto-gen/logs/v1"
 )
 
-type otlpExporter struct {
-	exporter *exporterImp
+type exporterImp struct {
+	// Input configuration.
+	config *Config
+
 	stopOnce sync.Once
+	w        sender
 }
 
-type exporterErrorCode int
-type exporterError struct {
-	code exporterErrorCode
-	msg  string
+type sender interface {
+	exportTrace(ctx context.Context, request *otlptrace.ExportTraceServiceRequest) error
+	exportMetrics(ctx context.Context, request *otlpmetrics.ExportMetricsServiceRequest) error
+	exportLogs(ctx context.Context, request *otlplogs.ExportLogServiceRequest) error
+	stop() error
 }
 
-var _ error = (*exporterError)(nil)
-
-func (e *exporterError) Error() string {
-	return e.msg
-}
-
-const (
-	_ exporterErrorCode = iota // skip 0
-	// errEndpointRequired indicates that this exporter was not provided with an endpoint in its config.
-	errEndpointRequired
-	// errAlreadyStopped indicates that the exporter was already stopped.
-	errAlreadyStopped
+var (
+	errTimeout    = errors.New("timeout")
+	errFatalError = errors.New("fatal error sending to server")
 )
 
-// NewTraceExporter creates an OTLP trace exporter.
-func NewTraceExporter(
-	ctx context.Context,
-	params component.ExporterCreateParams,
-	config configmodels.Exporter,
-) (component.TraceExporter, error) {
-	oce, err := createOTLPExporter(config)
-	if err != nil {
-		return nil, err
-	}
-	oexp, err := exporterhelper.NewTraceExporter(
-		config,
-		oce.pushTraceData,
-		exporterhelper.WithShutdown(oce.Shutdown))
-	if err != nil {
-		return nil, err
-	}
-
-	return oexp, nil
-}
-
-// NewMetricsExporter creates an OTLP metrics exporter.
-func NewMetricsExporter(
-	_ context.Context,
-	_ component.ExporterCreateParams,
-	config configmodels.Exporter,
-) (component.MetricsExporter, error) {
-	oce, err := createOTLPExporter(config)
-	if err != nil {
-		return nil, err
-	}
-	oexp, err := exporterhelper.NewMetricsExporter(
-		config,
-		oce.pushMetricsData,
-		exporterhelper.WithShutdown(oce.Shutdown),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	return oexp, nil
-}
-
-// NewLogExporter creates an OTLP log exporter.
-func NewLogExporter(
-	_ context.Context,
-	_ component.ExporterCreateParams,
-	config configmodels.Exporter,
-) (component.LogExporter, error) {
-	oce, err := createOTLPExporter(config)
-	if err != nil {
-		return nil, err
-	}
-	oexp, err := exporterhelper.NewLogsExporter(
-		config,
-		oce.pushLogData,
-		exporterhelper.WithShutdown(oce.Shutdown),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	return oexp, nil
-}
-
-// createOTLPExporter creates an OTLP exporter.
-func createOTLPExporter(config configmodels.Exporter) (*otlpExporter, error) {
-	oCfg := config.(*Config)
+// Crete new exporter and start it. The exporter will begin connecting but
+// this function may return before the connection is established.
+func newExporter(cfg configmodels.Exporter) (*exporterImp, error) {
+	oCfg := cfg.(*Config)
 
 	if oCfg.Endpoint == "" {
-		return nil, &exporterError{
-			code: errEndpointRequired,
-			msg:  "OTLP exporter config requires an Endpoint",
-		}
+		return nil, errors.New("OTLP exporter config requires an Endpoint")
 	}
 
-	exporter, serr := newExporter(oCfg)
-	if serr != nil {
-		return nil, fmt.Errorf("cannot configure OTLP exporter: %v", serr)
+	e := &exporterImp{}
+	e.config = oCfg
+	w, err := newGrpcSender(oCfg)
+	if err != nil {
+		return nil, err
 	}
-
-	oce := &otlpExporter{exporter: exporter}
-	return oce, nil
+	e.w = w
+	return e, nil
 }
 
-func (oce *otlpExporter) Shutdown(context.Context) error {
-	err := error(&exporterError{
-		code: errAlreadyStopped,
-		msg:  "OpenTelemetry exporter was already stopped.",
-	})
-	oce.stopOnce.Do(func() {
-		err = oce.exporter.stop()
+func (e *exporterImp) shutdown(context.Context) error {
+	err := componenterror.ErrAlreadyStopped
+	e.stopOnce.Do(func() {
+		// Close the connection.
+		err = e.w.stop()
 	})
 	return err
 }
 
-func (oce *otlpExporter) pushTraceData(ctx context.Context, td pdata.Traces) (int, error) {
+func (e *exporterImp) pushTraceData(ctx context.Context, td pdata.Traces) (int, error) {
 	request := &otlptrace.ExportTraceServiceRequest{
 		ResourceSpans: pdata.TracesToOtlp(td),
 	}
-	err := oce.exporter.exportTrace(ctx, request)
+	err := e.w.exportTrace(ctx, request)
 
 	if err != nil {
 		return td.SpanCount(), fmt.Errorf("failed to push trace data via OTLP exporter: %w", err)
@@ -163,12 +98,12 @@ func (oce *otlpExporter) pushTraceData(ctx context.Context, td pdata.Traces) (in
 	return 0, nil
 }
 
-func (oce *otlpExporter) pushMetricsData(ctx context.Context, md pdata.Metrics) (int, error) {
+func (e *exporterImp) pushMetricsData(ctx context.Context, md pdata.Metrics) (int, error) {
 	imd := pdatautil.MetricsToInternalMetrics(md)
 	request := &otlpmetrics.ExportMetricsServiceRequest{
 		ResourceMetrics: data.MetricDataToOtlp(imd),
 	}
-	err := oce.exporter.exportMetrics(ctx, request)
+	err := e.w.exportMetrics(ctx, request)
 
 	if err != nil {
 		return imd.MetricCount(), fmt.Errorf("failed to push metrics data via OTLP exporter: %w", err)
@@ -176,14 +111,190 @@ func (oce *otlpExporter) pushMetricsData(ctx context.Context, md pdata.Metrics) 
 	return 0, nil
 }
 
-func (oce *otlpExporter) pushLogData(ctx context.Context, logs data.Logs) (int, error) {
+func (e *exporterImp) pushLogData(ctx context.Context, logs data.Logs) (int, error) {
 	request := &otlplogs.ExportLogServiceRequest{
 		ResourceLogs: data.LogsToProto(logs),
 	}
-	err := oce.exporter.exportLogs(ctx, request)
+	err := e.w.exportLogs(ctx, request)
 
 	if err != nil {
 		return logs.LogRecordCount(), fmt.Errorf("failed to push log data via OTLP exporter: %w", err)
 	}
 	return 0, nil
+}
+
+type grpcSender struct {
+	// gRPC clients and connection.
+	traceExporter  otlptrace.TraceServiceClient
+	metricExporter otlpmetrics.MetricsServiceClient
+	logExporter    otlplogs.LogServiceClient
+	grpcClientConn *grpc.ClientConn
+	metadata       metadata.MD
+	waitForReady   bool
+}
+
+func newGrpcSender(config *Config) (sender, error) {
+	dialOpts, err := config.GRPCClientSettings.ToDialOptions()
+	if err != nil {
+		return nil, err
+	}
+
+	var clientConn *grpc.ClientConn
+	if clientConn, err = grpc.Dial(config.GRPCClientSettings.Endpoint, dialOpts...); err != nil {
+		return nil, err
+	}
+
+	gs := &grpcSender{
+		traceExporter:  otlptrace.NewTraceServiceClient(clientConn),
+		metricExporter: otlpmetrics.NewMetricsServiceClient(clientConn),
+		logExporter:    otlplogs.NewLogServiceClient(clientConn),
+		grpcClientConn: clientConn,
+		metadata:       metadata.New(config.GRPCClientSettings.Headers),
+		waitForReady:   config.GRPCClientSettings.WaitForReady,
+	}
+	return gs, nil
+}
+
+func (gs *grpcSender) stop() error {
+	return gs.grpcClientConn.Close()
+}
+
+func (gs *grpcSender) exportTrace(ctx context.Context, request *otlptrace.ExportTraceServiceRequest) error {
+	return exportRequest(gs.enhanceContext(ctx), func(ctx context.Context) error {
+		_, err := gs.traceExporter.Export(ctx, request, grpc.WaitForReady(gs.waitForReady))
+		return err
+	})
+}
+
+func (gs *grpcSender) exportMetrics(ctx context.Context, request *otlpmetrics.ExportMetricsServiceRequest) error {
+	return exportRequest(gs.enhanceContext(ctx), func(ctx context.Context) error {
+		_, err := gs.metricExporter.Export(ctx, request, grpc.WaitForReady(gs.waitForReady))
+		return err
+	})
+}
+
+func (gs *grpcSender) exportLogs(ctx context.Context, request *otlplogs.ExportLogServiceRequest) error {
+	return exportRequest(gs.enhanceContext(ctx), func(ctx context.Context) error {
+		_, err := gs.logExporter.Export(ctx, request, grpc.WaitForReady(gs.waitForReady))
+		return err
+	})
+}
+
+func (gs *grpcSender) enhanceContext(ctx context.Context) context.Context {
+	if gs.metadata.Len() > 0 {
+		return metadata.NewOutgoingContext(ctx, gs.metadata)
+	}
+	return ctx
+}
+
+// Send a trace or metrics request to the server. "perform" function is expected to make
+// the actual gRPC unary call that sends the request. This function implements the
+// common OTLP logic around request handling such as retries and throttling.
+func exportRequest(ctx context.Context, perform func(ctx context.Context) error) error {
+
+	expBackoff := backoff.NewExponentialBackOff()
+
+	// Spend max 15 mins on this operation. This is just a reasonable number that
+	// gives plenty of time for typical quick transient errors to resolve.
+	expBackoff.MaxElapsedTime = time.Minute * 15
+
+	for {
+		// Send to server.
+		err := perform(ctx)
+
+		if err == nil {
+			// Request is successful, we are done.
+			return nil
+		}
+
+		// We have an error, check gRPC status code.
+
+		status := status.Convert(err)
+
+		statusCode := status.Code()
+		if statusCode == codes.OK {
+			// Not really an error, still success.
+			return nil
+		}
+
+		// Now, this is this a real error.
+
+		if !shouldRetry(statusCode) {
+			// It is not a retryable error, we should not retry.
+			return errFatalError
+		}
+
+		// Need to retry.
+
+		// Check if server returned throttling information.
+		waitDuration := getThrottleDuration(status)
+		if waitDuration == 0 {
+			// No explicit throttle duration. Use exponential backoff strategy.
+			waitDuration = expBackoff.NextBackOff()
+			if waitDuration == backoff.Stop {
+				// We run out of max time allocated to this operation.
+				return errTimeout
+			}
+		}
+
+		// Wait until one of the conditions below triggers.
+		select {
+		case <-ctx.Done():
+			// This request is cancelled or timed out.
+			return errTimeout
+
+		case <-time.After(waitDuration):
+			// Time to try again.
+		}
+	}
+}
+
+func shouldRetry(code codes.Code) bool {
+	switch code {
+	case codes.OK:
+		// Success. This function should not be called for this code, the best we
+		// can do is tell the caller not to retry.
+		return false
+
+	case codes.Canceled,
+		codes.DeadlineExceeded,
+		codes.PermissionDenied,
+		codes.Unauthenticated,
+		codes.ResourceExhausted,
+		codes.Aborted,
+		codes.OutOfRange,
+		codes.Unavailable,
+		codes.DataLoss:
+		// These are retryable errors.
+		return true
+
+	case codes.Unknown,
+		codes.InvalidArgument,
+		codes.NotFound,
+		codes.AlreadyExists,
+		codes.FailedPrecondition,
+		codes.Unimplemented,
+		codes.Internal:
+		// These are fatal errors, don't retry.
+		return false
+
+	default:
+		// Don't retry on unknown codes.
+		return false
+	}
+}
+
+func getThrottleDuration(status *status.Status) time.Duration {
+	// See if throttling information is available.
+	for _, detail := range status.Details() {
+		switch t := detail.(type) {
+		case *errdetails.RetryInfo:
+			if t.RetryDelay.Seconds > 0 || t.RetryDelay.Nanos > 0 {
+				// We are throttled. Wait before retrying as requested by the server.
+				return time.Duration(t.RetryDelay.Seconds)*time.Second + time.Duration(t.RetryDelay.Nanos)*time.Nanosecond
+			}
+			return 0
+		}
+	}
+	return 0
 }
